requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pytest>=7.4.0
pytest-django>=4.5.0
playwright>=1.40.0
pytest-playwright>=0.4.3
gradio>=4.0.0
